# storage/api-docs-v1/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "One Dashboard API V1 Swagger Documentation",
        "version": "1"
    },
    "paths": {
...
        "/api/v1/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get list transactions ",
                "description": "Endpoint to get list transactions ",
                "operationId": "0ffe6bf7980513c9d322e1ba4d400ff2",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category ID"
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Product ID"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Code"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type ('IN' or 'OUT')"
                    },
                    {
                        "name": "total_item",
                        "in": "query",
                        "description": "Total Item"
                    },
                    {
                        "name": "total_quantity",
                        "in": "query",
                        "description": "Total Quantity"
                    },
                    {
                        "name": "total_price",
                        "in": "query",
                        "description": "Total Price"
                    },
                    {
                        "name": "customer_name",
                        "in": "query",
                        "description": "Customer Name"
                    },
                    {
                        "name": "customer_phone",
                        "in": "query",
                        "description": "Customer Phone"
                    },
                    {
                        "name": "customer_address",
                        "in": "query",
                        "description": "Customer Address"
                    },
                    {
                        "name": "note",
                        "in": "query",
                        "description": "Note"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "1 for Ascending -1 for Descending"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit (Default 10)"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Num Of Page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list transactions successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "pagination": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Get list transactions failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get list transactions failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create a new transaction",
                "description": "Create a new transaction",
                "operationId": "b0ee1d82e907b370897f7c2adde8a664",
                "requestBody": {
                    "description": "Data that needed to create a new transaction",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "note",
                                    "items"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "customer_address": {
                                        "type": "string",
                                        "example": "Jl. Raya No. 1"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Buy some items"
                                    },
                                    "items": {
                                        "description": "Array of transaction items",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "description": "Product ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "description": "Quantity of the item",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "note": {
                                                    "description": "Note of the item",
                                                    "type": "string",
                                                    "example": "Buy some items"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new transaction successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Register a new transaction successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Create a new transactions failed"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/transactions/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get detail transaction",
                "description": "Endpoint to get detail transaction",
                "operationId": "c9e072c8c645241005e971a691e8338e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get transaction successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get transaction successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}